move to internal/metrics dir
